reorganized files into better looking module system
