--- ecdsa.py
@@ -1,8 +1,0 @@
-#!/usr/bin/python3
-
-from keygen import keygen
-from sign import sign
-from verify import verify
-
-if __name__ == "__main__":
-    print("Use `ecdsa` as library module")
