+++ ecdsa/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/python3
+
+from .keygen import keygen
+from .sign import sign
+from .verify import verify
