+++ field/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/python3
+
+from utils import bit_count
+from .base_field import BaseField
+from .base_field_consts import P
+from .scalar_field import ScalarField
+from .scalar_field_consts import N, Gx, Gy
